requirements.txt
anyio==4.11.0
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
colorama==0.4.6
Django==5.2.7
djangorestframework==3.16.1
dotenv==0.9.9
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-django==4.11.1
python-dotenv==1.2.1
python-telegram-bot==22.5
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
